chore: change quantity datatype to integer

+++ src/main/java/com/fawry/store/dtos/StockRequestDto.java
@@ -6,6 +6,6 @@ import jakarta.validation.constraints.NotNull;
 public record StockRequestDto(
         @Digits(integer = 10, fraction = 0, message = "Quantity must be a number")
         @NotNull(message = "Quantity is required")
-        Long quantity
+        int quantity
 ) {
 }
